+++ management-api-server/src/main/java/com/datastax/mgmtapi/resources/NodeOpsResources.java
@@ -663,6 +663,48 @@ public class NodeOpsResources extends BaseResources {
         });
   }
 
+  @POST
+  @Path("/encryption/internode/truststore/reload")
+  @Produces(MediaType.TEXT_PLAIN)
+  @ApiResponse(
+      responseCode = "200",
+      description = "Truststore reloaded successfully",
+      content =
+          @Content(
+              mediaType = MediaType.TEXT_PLAIN,
+              schema = @Schema(implementation = String.class),
+              examples = @ExampleObject(value = "OK")))
+  @ApiResponse(
+      responseCode = "400",
+      description = "Unsupported Operation",
+      content =
+          @Content(
+              mediaType = MediaType.TEXT_PLAIN,
+              schema = @Schema(implementation = String.class),
+              examples =
+                  @ExampleObject(
+                      value = "Reloading the truststore manually is only possible on DSE")))
+  @Operation(summary = "reload truststore", operationId = "reloadTruststore")
+  public Response reloadTruststore() {
+    return handle(
+        () -> {
+          final String releaseVersion =
+              ResponseTools.getSingleRowStringResponse(
+                  app.dbUnixSocketFile, app.cqlService, CASSANDRA_VERSION_CQL_STRING);
+          if (!releaseVersion.startsWith("4.0.0.68") && !releaseVersion.startsWith("4.0.0.69")) {
+            // rebuilding search index is only possible on DSE
+            return Response.status(Response.Status.BAD_REQUEST)
+                .entity("Reloading the truststore manually is only possible on DSE")
+                .build();
+          }
+
+          app.cqlService.executeCql(
+              app.dbUnixSocketFile, "CALL NodeOps.reloadInternodeEncryptionTruststore()");
+
+          return Response.ok("OK").build();
+        });
+  }
+
   private static final String FQL_QUERY_RESPONSE_EXAMPLE =
       "{\n"
           + "    \"entity\": false,\n"